fix: correctly calculate indents when using "first" setting
When using the "first" setting to match indentation to the first token
of a list, the calculation was returning the total indent length instead
of the required multiples of indentation e.g. 20 instead of 10 if using
double spaced indentation. Fixed by dividing the total length by the
indent size.

Additionally ensures to match identation to the first token of a list
when the first token of an element begins with a punctuator. E.g.
```
// always valid
func('hello',
     'there')

// previously invalid
func('hello'
   , 'there')
```

Signed-off-by: Jacob Hull <jacob@planethull.com>